Fixed and added multiple controllers

+++ FRS/src/main/java/com/group/FRS/controller/PassengerController.java
@@ -30,50 +30,70 @@ public class PassengerController {
 
 	@Autowired
 	PassengerRepository passengerRepository;
-	
+
 	@Autowired
 	FlightRepository flightRepository;
-	
-	@GetMapping(path="/passengers")
-	public List<Passenger> getAllPassengers(){
+
+	@GetMapping(path = "/passengers")
+	public List<Passenger> getAllPassengers() {
 		return passengerRepository.findAll();
 	}
-	
-	@GetMapping(path="/passenger/{id}")
-	public Passenger getSinglePassenger(@PathVariable(value="id") Long passengerId) {
+
+	@GetMapping(path = "/passenger/{id}")
+	public Passenger getSinglePassenger(@PathVariable(value = "id") Long passengerId) {
 		return passengerRepository.findById(passengerId).orElse(null);
 	}
 
-	@PostMapping(path="/passenger")
-    public ResponseEntity<Passenger> create(@RequestBody Passenger passenger){
+	@PostMapping(path = "/passenger")
+	public ResponseEntity<Passenger> create(@RequestBody Passenger passenger) {
+		return ResponseEntity.ok(passengerRepository.save(createOneToMany(passenger)));
+	}
+	
+	public static Passenger createOneToMany(Passenger passenger) {
+		Passenger pC = new Passenger(); 
+		pC.setAge(passenger.getAge());
+		pC.setFirstName(passenger.getFirstName());
+		pC.setLastName(passenger.getLastName());
+		pC.setGender(passenger.getGender());
+		pC.setSeatNo(passenger.getSeatNo());
+		pC.setBookingDate(passenger.getBookingDate());
+		for(PassengerSchedule ps: passenger.getPassengerSchedules()) {
+			pC.addPassengerSchedule(ps);
+		}
+		return pC;
+	}
+	
+	
+
+	@PutMapping(path = "/passenger/{id}")
+	public ResponseEntity<Passenger> updatePassenger(@PathVariable(value = "id") Long passengerId,
+			@Valid @RequestBody Passenger passengerDetails) {
+		Passenger passenger = passengerRepository.findById(passengerId).orElse(null);
+		passenger.setAge(passengerDetails.getAge());
+		passenger.setFirstName(passengerDetails.getFirstName());
+		passenger.setLastName(passengerDetails.getLastName());
+		passenger.setGender(passengerDetails.getGender());
+		passenger.setSeatNo(passengerDetails.getSeatNo());
+		passenger.setBookingDate(passengerDetails.getBookingDate());
 		return ResponseEntity.ok(passengerRepository.save(passenger));
-    }
-	
-	@PutMapping(path="/passenger/{id}")
-	   public ResponseEntity<Passenger> updatePassenger(@PathVariable(value = "id") Long passengerId,
-	                                              @Valid @RequestBody Passenger passengerDetails) {
-	       Passenger passenger = passengerRepository.findById( passengerId).orElse(null);
-	       passenger.setAge(passengerDetails.getAge());
-	       passenger.setFirstName(passengerDetails.getFirstName());
-	       passenger.setLastName(passengerDetails.getLastName());
-	       passenger.setGender(passengerDetails.getGender());
-	       passenger.setSeatNo(passengerDetails.getSeatNo());
-	       passenger.setBookingDate(passengerDetails.getBookingDate());
-	       return ResponseEntity.ok(passengerRepository.save(passenger));
-	   }
-	
-	   @PutMapping("/passenger/{id1}/flight/{id2}")
-	   	public ResponseEntity<Passenger> connect(@PathVariable(value = "id") Long passengerId,
-	   			@PathVariable(value = "id2") Long flightId) {
-	   		Passenger pass = passengerRepository.findById(passengerId).orElse(null);
-	   		Flight flight = flightRepository.findById(flightId).orElse(null);
-	   		pass.setFlight(flight);
-	   		return ResponseEntity.ok(passengerRepository.save(pass));
-	   	}
-	
-	@DeleteMapping(path="/passenger/{id}")
-    public void delete(@PathVariable("id") Long id) {
-         passengerRepository.deleteById(id);
-    }
-	
+	}
+
+	@PutMapping("/passenger/{id1}/flight/{id2}")
+	public ResponseEntity<Passenger> connect(@PathVariable(value = "id") Long passengerId,
+			@PathVariable(value = "id2") Long flightId) {
+		Passenger pass = passengerRepository.findById(passengerId).orElse(null);
+		Flight flight = flightRepository.findById(flightId).orElse(null);
+		pass.setFlight(flight);
+		return ResponseEntity.ok(passengerRepository.save(pass));
+	}
+
+	@DeleteMapping("/passenger/{id}")
+	public ResponseEntity<Passenger> delete(@PathVariable("id") Long id) {
+		if (passengerRepository.findById(id).orElse(null) != null) {
+			passengerRepository.deleteById(id);
+			return ResponseEntity.ok().body(null);
+		}
+		return ResponseEntity.notFound().build();
+	}
+
 }
